--- Example/Podfile
@@ -1,11 +1,0 @@
-source 'https://github.com/CocoaPods/Specs.git'
-
-use_frameworks!
-platform :ios, '12.0'
-
-target 'JNPhoneNumberView_Example' do
-  pod 'JNPhoneNumberView', :path => '../'
-
-  pod 'libPhoneNumber-iOS'
-end
-
