--- JNPhoneNumberView/Classes/JNCountryCodePickerViewController/ViewController/JNCountryPickerViewController.xib
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14490.70" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14490.49"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="JNCountryPickerViewController" customModule="JNPhoneNumberView" customModuleProvider="target">
-            <connections>
-                <outlet property="searchBar" destination="jEe-g2-kwl" id="gNn-Ds-ol7"/>
-                <outlet property="tableView" destination="XAe-Pd-PD2" id="YG3-7l-U7t"/>
-                <outlet property="view" destination="XtZ-bb-rBp" id="fbv-Hu-Owh"/>
-            </connections>
-        </placeholder>
-        <view contentMode="scaleToFill" id="XtZ-bb-rBp">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <subviews>
-                <searchBar contentMode="redraw" translatesAutoresizingMaskIntoConstraints="NO" id="jEe-g2-kwl">
-                    <rect key="frame" x="0.0" y="20" width="375" height="56"/>
-                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                    <textInputTraits key="textInputTraits" spellCheckingType="yes"/>
-                </searchBar>
-                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="none" rowHeight="44" sectionHeaderHeight="1" sectionFooterHeight="1" translatesAutoresizingMaskIntoConstraints="NO" id="XAe-Pd-PD2">
-                    <rect key="frame" x="0.0" y="76" width="375" height="591"/>
-                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                    <connections>
-                        <outlet property="dataSource" destination="-1" id="Irp-Zr-C86"/>
-                        <outlet property="delegate" destination="-1" id="qj9-xI-Y1e"/>
-                    </connections>
-                </tableView>
-            </subviews>
-            <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-            <constraints>
-                <constraint firstItem="pbl-jX-wXc" firstAttribute="bottom" secondItem="XAe-Pd-PD2" secondAttribute="bottom" id="IpN-6l-CCu"/>
-                <constraint firstAttribute="trailing" secondItem="XAe-Pd-PD2" secondAttribute="trailing" id="S2F-z9-PZa"/>
-                <constraint firstItem="jEe-g2-kwl" firstAttribute="top" secondItem="pbl-jX-wXc" secondAttribute="top" id="YST-1b-Uuv"/>
-                <constraint firstAttribute="trailing" secondItem="jEe-g2-kwl" secondAttribute="trailing" id="fFF-KA-A1V"/>
-                <constraint firstItem="XAe-Pd-PD2" firstAttribute="top" secondItem="jEe-g2-kwl" secondAttribute="bottom" id="kD8-eZ-tzP"/>
-                <constraint firstItem="XAe-Pd-PD2" firstAttribute="leading" secondItem="XtZ-bb-rBp" secondAttribute="leading" id="tLl-fg-nm8"/>
-                <constraint firstItem="jEe-g2-kwl" firstAttribute="leading" secondItem="XtZ-bb-rBp" secondAttribute="leading" id="wm6-mJ-RLS"/>
-            </constraints>
-            <viewLayoutGuide key="safeArea" id="pbl-jX-wXc"/>
-        </view>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-    </objects>
-</document>
