+++ README.md
@@ -7,19 +7,15 @@
 
 ## Requirements
 
-- iOS 12.0+ / macOS 14.4+
+- iOS 12.0+ / macOS 14.0+
 - Xcode 15.3+
-- Swift 5+
+- Swift 5.10+
 
 
 ## Installation
 
-JNPhoneNumberView is available through [CocoaPods](https://cocoapods.org). To install
-it, simply add the following line to your Podfile:
+JNPhoneNumberView support SwiftPM. To use SwiftPM, you should use Xcode 15.3 to open your project. Click `File` -> `Swift Packages` -> `Add Package Dependency`, enter [JNPhoneNumberView repo's URL](https://github.com/JNDisrupter/JNPhoneNumberView.git)
 
-```ruby
-pod 'JNPhoneNumberView'
-```
 ## Example
 
 To run the example project, clone the repo, and run `pod install` from the Example directory first.
