+++ Sources/JNPhoneNumberView/Entites/GeneralResource/GeneralResource.swift
