+++ Sources/JNPhoneNumberView/Entites/Representables/GeneralResourceRepresentable/GeneralResourceTableViewCellRepresentable.swift
