+++ Sources/JNPhoneNumberView/Resources/Assets/Assets.xcassets/Contents.json
