+++ Sources/JNPhoneNumberView/Resources/Assets/Assets.xcassets/icons/Contents.json
