+++ Sources/JNPhoneNumberView/Resources/Assets/Assets.xcassets/icons/EmptySearchResult.imageset/Contents.json
