+++ Sources/JNPhoneNumberView/Resources/Assets/Assets.xcassets/icons/EmptySearchResult.imageset/EmptySearchResult.pdf
