+++ Sources/JNPhoneNumberView/Resources/Assets/Assets.xcassets/icons/closeImageButton.imageset/Contents.json
