+++ Sources/JNPhoneNumberView/Resources/Assets/Assets.xcassets/icons/selectedPickerImage.imageset/Contents.json
