+++ Sources/JNPhoneNumberView/Resources/Assets/Assets.xcassets/icons/selectedPickerImage.imageset/selectedPickerImage.pdf
